+++ app/tests/addProject.page.js
@@ -27,13 +27,17 @@ class AddProjectPage {
     await testController.typeText('#project-form-mail-address', project.mailAddress.address);
     await testController.typeText('#project-form-mail-address2', project.mailAddress.address2);
     await testController.typeText('#project-form-mail-city', project.mailAddress.city);
-    await testController.typeText('#project-form-mail-state', project.mailAddress.state);
+    const selectState = Selector('#project-form-mail-state');
+    await testController.click(selectState);
+    await testController.click(selectState.find('option[value="AL"]'));
     await testController.typeText('#project-form-mail-zip', project.mailAddress.zip);
     // await testController.typeText('project-form-mailAddress-country', project.mailAddress.country);
     await testController.typeText('#project-form-ship-address', project.shipAddress.address);
     await testController.typeText('#project-form-ship-address2', project.shipAddress.address2);
     await testController.typeText('#project-form-ship-city', project.shipAddress.city);
-    await testController.typeText('#project-form-ship-state', project.shipAddress.state);
+    const selectState2 = Selector('#project-form-ship-state');
+    await testController.click(selectState2);
+    await testController.click(selectState2.find('option[value="AL"]'));
     await testController.typeText('#project-form-ship-zip', project.shipAddress.zip);
     // await testController.typeText('project-form-shipAddress-country', project.shipAddress.country);
     await testController.click('#project-form-submit input.btn.btn-primary');
